refactor a calculate_reachable_functions method
